requirements.txt
asgiref==3.8.1
cachetools==5.5.0
Django==4.2.16
google-auth==2.35.0
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
requests
